validations TELEF_PERSO

+++ CRUDS.sql
@@ -11,7 +11,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM PAIS WHERE Cod_Pais = pCodPais) = 0) THEN
-				IF (pNombrePais != '') THEN
+				IF (pNombrePais != '' AND pNombrePais IS NOT NULL) THEN
 					INSERT INTO PAIS(Cod_Pais,Nombre_Pais)
 					VALUES(pCodPais,pNombrePais);
 				ELSE
@@ -78,7 +78,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM PROVINCIA WHERE Cod_Provincia = pCodProvincia) = 0) THEN
-				IF (pNombreProvincia != '') THEN
+				IF (pNombreProvincia != ''AND pNombreProvincia IS NOT NULL) THEN
 					IF ((SELECT COUNT(*) FROM PAIS WHERE Cod_Pais = pCodPais) > 0) THEN 
 						INSERT INTO PROVINCIA(Cod_Provincia, Nombre_Provincia, Cod_Pais)
 						VALUES(pCodProvincia, pNombreProvincia, pCodPais);
@@ -149,7 +149,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM CIUDAD WHERE Cod_Ciudad = pCodCiudad) = 0) THEN
-				IF (pNombreCiudad != '') THEN
+				IF (pNombreCiudad != '' AND pNombreCiudad IS NOT NULL) THEN
 					IF ((SELECT COUNT(*) FROM PROVINCIA WHERE Cod_Provincia = pCodProvincia) > 0) THEN 
 						INSERT INTO CIUDAD(Cod_Ciudad, Nombre_Ciudad, Cod_Provincia)
 						VALUES(pCodCiudad, pNombreCiudad, pCodProvincia);
@@ -229,7 +229,7 @@ BEGIN
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM SUCURSAL WHERE Cod_Sucursal = pCodSucursal) = 0) THEN
-				IF (pNombreCiudad != '') THEN
+				IF (pNombreSucursal != '' AND pNombreSucursal IS NOT NULL) THEN
 					IF ((SELECT COUNT(*) FROM CIUDAD WHERE Cod_Ciudad = pCodCiudad) > 0) THEN 
 						INSERT INTO SUCURSAL(Cod_Sucursal, Nombre_Sucursal, Cod_Ciudad)
 						VALUES(pCodSucursal, pNombreSucursal, pCodCiudad);
@@ -305,7 +305,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM SEXO WHERE Cod_Sexo = pCodSexo) = 0) THEN
-				IF (pDescripSexo != '') THEN
+				IF (pDescripSexo != '' AND pDescripSexo IS NOT NULL) THEN
 					INSERT INTO SEXO(Cod_Sexo, Descrip_Sexo)
 					VALUES(pCodSexo, pDescripSexo);
 				ELSE
@@ -400,7 +400,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM PUESTO_LABORAL WHERE Cod_Puesto_Laboral = pCodPuLaboral) = 0) THEN
-				IF (pDescPuLaboral != '') THEN
+				IF (pDescPuLaboral != '' AND pDescPuLaboral IS NOT NULL) THEN
 					IF ((pSalarioMe IS NOT NULL) AND (pSalarioMe >= 0 )) THEN
 						INSERT INTO PUESTO_LABORAL(Cod_Puesto_Laboral, Descrip_Pues_Lab, Salario_Mensual)
 						VALUES(pCodPuLaboral, pDescPuLaboral, pSalarioMe);
@@ -504,7 +504,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM TIPO_USUARIO WHERE Cod_Tipo_Usuario = pCodTipUsu) = 0) THEN
-				IF (pDescTipUsu != '') THEN
+				IF (pDescTipUsu != '' AND pDescTipUsu IS NOT NULL) THEN
 					INSERT INTO TIPO_USUARIO(Cod_Tipo_Usuario, Descrip_TipUsu)
 					VALUES(pCodTipUsu, pDescTipUsu);
 				ELSE
@@ -650,7 +650,7 @@ DECLARE msgError VARCHAR(70) DEFAULT '';
     
 		IF (pOperacion = 'CREATE') THEN
 			IF ((SELECT COUNT(*) FROM IMPUESTO WHERE Cod_Impuesto = pCodImpu) = 0) THEN
-				IF (pDescImpu != '') THEN
+				IF (pDescImpu != ''  AND pDescImpu IS NOT NULL) THEN
 					IF ((pPorcImpu IS NOT NULL) AND (pPorcImpu >= 0 )) THEN
 						INSERT INTO IMPUESTO(Cod_Impuesto, Porcentaje_Imp, Descrip_Imp)
 						VALUES(pCodImpu, pPorcImpu, pDescImpu);
@@ -805,26 +805,64 @@ END;
 delimiter //
 CREATE PROCEDURE CRUD_TELEFPERSO(pIdentPers INT, pNumTel VARCHAR(15), pContTelef INT, pOperacion VARCHAR(10))
 BEGIN
-	IF (pOperacion = 'CREATE') THEN
-		INSERT INTO TELEF_PERSO(Identificacion_Per, Num_Telef)
-		VALUES(pIdentPers, pNumTel);
-	END IF;
-
-	IF (pOperacion = 'READ') THEN
-		SELECT Identificacion_Per, Num_Telef, Contador_Telef
-		FROM TELEF_PERSO
-		WHERE Identificacion_Per = pIdentPers;
-  END IF;
-
-	IF (pOperacion = 'UPDATE')  THEN
-		UPDATE TELEF_PERSO
-		SET Num_Telef = IFNULL(pNumTel,Num_Telef)
-		WHERE Identificacion_Per = pIdentPers AND Contador_Telef=pContTelef;
-	END IF;
+DECLARE msgError VARCHAR(70) DEFAULT '';
+	IF (pIdentPers IS NOT NULL) THEN
     
-	IF (pOperacion = 'DELETE') THEN
-		DELETE FROM TELEF_PERSO
-		WHERE Identificacion_Per = pIdentPers;
+		IF (pOperacion = 'CREATE') THEN
+			IF (pNumTel != '' AND pNumTel IS NOT NULL) THEN
+				INSERT INTO TELEF_PERSO(Identificacion_Per, Num_Telef)
+				VALUES(pIdentPers, pNumTel);
+			ELSE
+				SET msgError = 'El numero de telefono está vacío';
+				SELECT msgError;
+			END IF;
+		END IF;
+
+		IF (pOperacion = 'READ') THEN
+			IF ((SELECT COUNT(*) FROM TELEF_PERSO WHERE Identificacion_Per = pIdentPers) > 0) THEN
+				SELECT Identificacion_Per, Num_Telef, Contador_Telef
+				FROM TELEF_PERSO
+				WHERE Identificacion_Per = pIdentPers;
+			ELSE
+				SET msgError = 'La persona no tiene numeros de telefono asociados';
+				SELECT msgError;
+			END IF;
+		END IF;
+
+		IF (pOperacion = 'UPDATE')  THEN
+			IF ((SELECT COUNT(*) FROM TELEF_PERSO WHERE Identificacion_Per = pIdentPers) > 0) THEN
+				IF ((SELECT COUNT(*) FROM TELEF_PERSO WHERE Contador_Telef = pContTelef) > 0) THEN
+					UPDATE TELEF_PERSO
+					SET Num_Telef = IFNULL(pNumTel, Num_Telef)
+					WHERE Identificacion_Per = pIdentPers AND Contador_Telef = pContTelef;
+				ELSE
+					SET msgError = 'El ID del telefono no existe, imposible actualizar datos';
+					SELECT msgError;
+				END IF;
+            ELSE
+				SET msgError = 'La persona no tiene numeros de telefono asociados';
+				SELECT msgError;
+			END IF;
+		END IF;
+        
+		IF (pOperacion = 'DELETE') THEN
+			IF ((SELECT COUNT(*) FROM TELEF_PERSO WHERE Identificacion_Per = pIdentPers) > 0) THEN
+				IF ((SELECT COUNT(*) FROM TELEF_PERSO WHERE Contador_Telef = pContTelef) = 0) THEN
+					DELETE FROM TELEF_PERSO
+					WHERE Identificacion_Per = pIdentPers AND Contador_Telef = pContTelef;
+				ELSE
+					SET msgError = 'No se puede eliminar, el ID del telefono no existe';
+					SELECT msgError;
+				END IF;
+			ELSE
+				SET msgError = 'La persona no tiene numeros de telefono asociados';
+				SELECT msgError;
+			END IF;
+		END IF;
+        
+	ELSE
+		SET msgError = 'El numero de cedula es vacío';
+        SELECT msgError;
 	END IF;
 END;
 //
